Refactor: 포토부스와 포토부스 브랜드 모듈 분리
- 포토부스에서 브랜드를 모듈로 분리

Related to #27
