+++ src/brand/brand.service.spec.ts
@@ -1,0 +1,18 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { BrandService } from './brand.service';
+
+describe('BrandService', () => {
+  let service: BrandService;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      providers: [BrandService],
+    }).compile();
+
+    service = module.get<BrandService>(BrandService);
+  });
+
+  it('should be defined', () => {
+    expect(service).toBeDefined();
+  });
+});
