+++ src/event/event.service.spec.ts
@@ -6,7 +6,7 @@ import { PhotoBoothService } from '../photo-booth/photo-booth.service';
 import { Events } from './entity/event.entity';
 import { GetEventListDto } from './dto/get-event-list.dto';
 import { NotFoundException } from '@nestjs/common';
-import { PhotoBoothBrand } from '../photo-booth/entity/photo-booth-brand.entity';
+import { PhotoBoothBrand } from '../brand/entity/brand.entity';
 import { Hashtag } from '../hashtag/entity/hashtag.entity';
 
 class MockEventRepository {
