+++ src/photo-booth/dto/get-photo-booth-detail.dto.ts
@@ -1,10 +1,6 @@
 import { Expose } from 'class-transformer';
 import { ApiProperty } from '@nestjs/swagger';
-import {
-  GetBoothBrandListDto,
-  GetPhotoBoothListDto,
-} from './get-photo-booth-list.dto';
-import { BrandImage } from '../entity/photo-booth-brand-image.entity';
+import { GetPhotoBoothListDto } from './get-photo-booth-list.dto';
 
 export class GetPhotoBoothDetailDto extends GetPhotoBoothListDto {
   @ApiProperty({
@@ -32,40 +28,5 @@ export class GetPhotoBoothDetailDto extends GetPhotoBoothListDto {
   @Expose()
   get operationTime(): string | null {
     return this._operationTime;
-  }
-}
-
-export class GetBoothBrandDetailDto extends GetBoothBrandListDto {
-  @ApiProperty({
-    description: '포토부스의 업체 설명',
-  })
-  @Expose()
-  get description(): string | null {
-    return this._description;
-  }
-
-  @ApiProperty({
-    description: '포토부스의 업체 관련 홈페이지 주소',
-  })
-  @Expose()
-  get photoBoothUrl(): string | null {
-    return this._photoBoothUrl;
-  }
-
-  @ApiProperty({
-    description: '포토부스의 업체 이벤트 허용 여부',
-    example: true,
-  })
-  @Expose()
-  get isEvent(): boolean {
-    return this._isEvent;
-  }
-
-  @ApiProperty({
-    description: '포토부스의 업체 이벤트 허용 여부',
-  })
-  @Expose()
-  get images(): BrandImage[] {
-    return this._brandImages;
   }
 }
