+++ src/photo-booth/dto/get-photo-booth-list.dto.ts
@@ -1,10 +1,8 @@
 import { Exclude, Expose } from 'class-transformer';
 import { ApiProperty } from '@nestjs/swagger';
-import { PhotoBoothBrand } from '../entity/photo-booth-brand.entity';
-import { BrandHashtag } from '../../hashtag/entity/brand-hashtag.entity';
+import { PhotoBoothBrand } from '../../brand/entity/brand.entity';
 import { PhotoBooth } from '../entity/photo-booth.entity';
 import { HiddenPhotoBooth } from '../entity/photo-booth-hidden.entity';
-import { BrandImage } from '../entity/photo-booth-brand-image.entity';
 
 export class GetPhotoBoothListDto {
   @Exclude() readonly _id: string;
@@ -74,60 +72,5 @@ export class GetPhotoBoothListDto {
   @Expose()
   get brandName(): string {
     return this._photoBoothBrand.name;
-  }
-}
-
-export class GetBoothBrandListDto {
-  @Exclude() readonly _id: number;
-  @Exclude() readonly _name: string;
-  @Exclude() readonly _mainThumbnailImageUrl: string;
-  @Exclude() readonly _description: string;
-  @Exclude() readonly _photoBoothUrl: string;
-  @Exclude() readonly _isEvent: boolean;
-  @Exclude() readonly _brandHashtags: BrandHashtag[];
-  @Exclude() readonly _brandImages: BrandImage[];
-
-  constructor(data: PhotoBoothBrand) {
-    Object.keys(data).forEach((key) => (this[`_${key}`] = data[key]));
-  }
-
-  @ApiProperty({ description: '포토부스 업체의 id 값' })
-  @Expose()
-  get id(): number {
-    return this._id;
-  }
-
-  @ApiProperty({
-    description: '포토부스의 업체명',
-    example: '하루필름',
-  })
-  @Expose()
-  get name(): string {
-    return this._name;
-  }
-
-  @ApiProperty({
-    description: '포토부스의 업체 이벤트 허용 여부',
-    example: true,
-  })
-  @Expose()
-  get isEvent(): boolean {
-    return this._isEvent;
-  }
-
-  @ApiProperty({
-    description: '포토부스의 대표 이미지',
-  })
-  @Expose()
-  get mainThumbnailImageUrl(): string | null {
-    return this._mainThumbnailImageUrl;
-  }
-
-  @ApiProperty({
-    description: '포토부스의 해시태그',
-  })
-  @Expose()
-  get hashtags(): BrandHashtag[] {
-    return this._brandHashtags;
   }
 }
