+++ src/photo-booth/photo-booth.service.spec.ts
@@ -2,11 +2,11 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { PhotoBoothService } from './photo-booth.service';
 import { PhotoBoothRepository } from './repository/photo-booth.repository';
 import { HiddenBoothRepository } from './repository/photo-booth-hidden.repository';
-import { PhotoBoothBrandRepository } from './repository/photo-booth-brand.repository';
+import { PhotoBoothBrandRepository } from '../brand/repository/brand.repository';
 import { PhotoBooth } from './entity/photo-booth.entity';
 import { GetPhotoBoothListDto } from './dto/get-photo-booth-list.dto';
 import { ConflictException, NotFoundException } from '@nestjs/common';
-import { PhotoBoothBrand } from './entity/photo-booth-brand.entity';
+import { PhotoBoothBrand } from '../brand/entity/brand.entity';
 import { HiddenPhotoBooth } from './entity/photo-booth-hidden.entity';
 import { HashtagService } from '../hashtag/hashtag.service';
 import { Hashtag } from '../hashtag/entity/hashtag.entity';
