+++ src/photo-booth/photo-booth.service.ts
@@ -2,32 +2,20 @@ import { ConflictException, Injectable } from '@nestjs/common';
 import { PhotoBoothRepository } from './repository/photo-booth.repository';
 import { HiddenBoothRepository } from './repository/photo-booth-hidden.repository';
 import { PhotoBooth } from './entity/photo-booth.entity';
-import { PaginationProps } from 'src/common/dto/get-pagination-query.dto';
-import {
-  FindBoothOptionProps,
-  FindBrandOptionProps,
-} from './dto/get-photo-booth-query.dto';
+import { PaginationProps } from '../common/dto/get-pagination-query.dto';
+import { FindBoothOptionProps } from './dto/get-photo-booth-query.dto';
 import { HiddenPhotoBooth } from './entity/photo-booth-hidden.entity';
-import {
-  BrandUpdateProps,
-  PhotoBoothUpdateProps,
-} from './dto/patch-photo-booth.dto';
-import { PhotoBoothBrandRepository } from './repository/photo-booth-brand.repository';
-import { PhotoBoothBrand } from './entity/photo-booth-brand.entity';
+import { PhotoBoothUpdateProps } from './dto/patch-photo-booth.dto';
 import { MoveToOpenBoothProps } from './dto/put-photo-booth.dto';
-import { HashtagService } from '../hashtag/hashtag.service';
-import { BrandCreateProps } from './dto/post-photo-booth.dto';
 import { ResponseEntity } from '../common/entity/response.entity';
-import { BrandHashtag } from '../hashtag/entity/brand-hashtag.entity';
-import { Hashtag } from '../hashtag/entity/hashtag.entity';
+import { BrandService } from '../brand/brand.service';
 
 @Injectable()
 export class PhotoBoothService {
   constructor(
-    private readonly hashtagService: HashtagService,
     private readonly photoBoothRepository: PhotoBoothRepository,
     private readonly hiddenBoothRepository: HiddenBoothRepository,
-    private readonly photoBoothBrandRepository: PhotoBoothBrandRepository,
+    private readonly brandService: BrandService,
   ) {}
 
   /**
@@ -67,9 +55,9 @@ export class PhotoBoothService {
     updateProps: PhotoBoothUpdateProps,
   ): Promise<boolean> {
     await this.findOneOpenBooth(id);
-    const photoBoothBrand = await this.findOneBrandBy({
-      name: updateProps.brandName,
-    });
+    const photoBoothBrand = await this.brandService.findOneBrandByName(
+      updateProps.brandName,
+    );
 
     await this.photoBoothRepository.save({
       id,
@@ -157,10 +145,9 @@ export class PhotoBoothService {
       throw new ConflictException('이미 포토부스가 존재합니다.');
     }
 
-    const photoBoothBrand = await this.findOneBrandBy({
-      name: moveProps.brandName,
-    });
-
+    const photoBoothBrand = await this.brandService.findOneBrandByName(
+      moveProps.brandName,
+    );
     await this.deleteHiddenBooth(id);
     await this.photoBoothRepository.save({ id, photoBoothBrand, ...moveProps });
 
@@ -176,87 +163,5 @@ export class PhotoBoothService {
     await this.hiddenBoothRepository.save({ id, preprocessedAt: new Date() });
 
     return true;
-  }
-
-  /**
-   * @param pageProps - pagination (항목수, 페이지)
-   * @param query - request query string (업체명, 이벤트 허용 여부)
-   * @desc - 쿼리 파라미터에 맞는 포토부스 업체 반환
-   *       - 쿼리 옵션이 없으면 전체 포토부스 업체 반환
-   */
-  async findBrandByQueryParam(
-    pageProps: PaginationProps,
-    query: FindBrandOptionProps,
-  ): Promise<[PhotoBoothBrand[], number]> {
-    return this.photoBoothBrandRepository.findBrandByOptionAndCount(
-      PhotoBoothBrand.of(query),
-      pageProps,
-    );
-  }
-
-  /**
-   * @param props - 포토부스 업체의 id, 이름
-   * @desc 포토부스 업체명으로 업체에 대한 엔티티 반환
-   */
-  findOneBrandBy(props: FindBrandOptionProps): Promise<PhotoBoothBrand> {
-    // const findBrand = await this.photoBoothBrandRepository.findOneBrand(
-    //   PhotoBoothBrand.of(props),
-    // );
-    return this.photoBoothBrandRepository.findOneBrand(
-      PhotoBoothBrand.of(props),
-    );
-  }
-
-  /**
-   * @param createProps - 브랜드 생성 속성들
-   * @desc - 해시태그 생성
-   *       - 브랜드 생성
-   *       - 해시태그와 브랜드 연결
-   *       - 이벤트 여부, 업체명, 대표이미지, 해시태그들
-   */
-  async createBrandWithHastags(
-    createProps: BrandCreateProps,
-  ): Promise<boolean> {
-    const brandHashtags = await this.brandHashtags(createProps.hashtags);
-
-    await this.photoBoothBrandRepository.save({
-      brandHashtags,
-      ...createProps,
-    });
-
-    return true;
-  }
-
-  /**
-   * @param id          - 포토부스 업체에 대한 id
-   * @param updateProps - 수정이 필요한 데이터 일부
-   * @desc - 포토부스 업체의 이름, 설명, 업체 홈페이지 url, 해시태그, 대표이미지 수정
-   *       - 업체명, 설명, 홈페이지 주소, 대표이미지, 이벤트 여부, 해시태그
-   *       - 포토부스 업체 이미지를 여러장 수정
-   */
-  async updateBrandWithHastags(
-    id: number,
-    updateProps: BrandUpdateProps,
-  ): Promise<boolean> {
-    await this.findOneBrandBy({ id });
-
-    const brandHashtags = await this.brandHashtags(updateProps.hashtags);
-
-    await this.photoBoothBrandRepository.save({
-      id,
-      brandHashtags,
-      ...updateProps,
-    });
-
-    return true;
-  }
-
-  /**
-   * @param hashtags - 포토부스 업체의 생성 및 수정에 필요한 해시태그 엔티티
-   * @desc  - 포토부스 업체관련 해시태그 가져오기
-   */
-  private async brandHashtags(hashtags: Hashtag[]): Promise<BrandHashtag[]> {
-    const createdHashtags = await this.hashtagService.createHashtags(hashtags);
-    return createdHashtags.map((hashtag) => BrandHashtag.create(hashtag));
   }
 }
